Include more patches
Sometimes the most important stuff to know about is something that can't
merge.

* Patches routinely have a hold on them when they're looking for one
  more review.
* Optional CI jobs routinely fail.
* If a patch has an invalid bug label then it needs *more* attention,
  not less.

The number of simultaneous open reviews is not large and the labels are
displayed, so it is easy enough to mentally filter out PRs that have
various labels, but much harder to discover PRs that don't appear in the
list.